refactor: Rename input::engine namespace.
