+++ libopenage/input/controller/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_subdirectory("camera")
-add_subdirectory("engine")
+add_subdirectory("game")
