+++ libopenage/input/controller/game/CMakeLists.txt
