+++ libopenage/input/controller/game/binding.cpp
@@ -2,6 +2,6 @@
 
 #include "binding.h"
 
-namespace openage::input::engine {
+namespace openage::input::game {
 
 } // namespace openage::input::engine
