+++ libopenage/input/input_context.cpp
@@ -15,7 +15,7 @@ const std::string &InputContext::get_id() {
 	return this->id;
 }
 
-void InputContext::set_engine_bindings(const std::shared_ptr<engine::BindingContext> &bindings) {
+void InputContext::set_engine_bindings(const std::shared_ptr<game::BindingContext> &bindings) {
 	this->engine_bindings = bindings;
 }
 
@@ -23,7 +23,7 @@ void InputContext::set_camera_bindings(const std::shared_ptr<camera::BindingCont
 	this->camera_bindings = bindings;
 }
 
-const std::shared_ptr<engine::BindingContext> &InputContext::get_engine_bindings() {
+const std::shared_ptr<game::BindingContext> &InputContext::get_engine_bindings() {
 	return this->engine_bindings;
 }
