+++ libopenage/presenter/presenter.cpp
@@ -11,8 +11,8 @@
 #include "gamestate/engine.h"
 #include "input/controller/camera/binding_context.h"
 #include "input/controller/camera/controller.h"
-#include "input/controller/engine/binding_context.h"
-#include "input/controller/engine/controller.h"
+#include "input/controller/game/binding_context.h"
+#include "input/controller/game/controller.h"
 #include "input/input_context.h"
 #include "input/input_manager.h"
 #include "log/log.h"
@@ -201,10 +201,10 @@ void Presenter::init_input() {
 	// setup simulation controls
 	if (this->simulation) {
 		// TODO: Remove hardcoding
-		auto engine_controller = std::make_shared<input::engine::Controller>(
+		auto engine_controller = std::make_shared<input::game::Controller>(
 			std::unordered_set<size_t>{0, 1, 2, 3}, 0);
-		auto engine_context = std::make_shared<input::engine::BindingContext>();
-		input::engine::setup_defaults(engine_context, this->time_loop, this->simulation, this->camera);
+		auto engine_context = std::make_shared<input::game::BindingContext>();
+		input::game::setup_defaults(engine_context, this->time_loop, this->simulation, this->camera);
 		this->input_manager->set_engine_controller(engine_controller);
 		input_ctx->set_engine_bindings(engine_context);
 	}
